Break out classes/interfaces to separate files
