+++ src/ume-app-estateservice/Umea.se.EstateService.Logic/Handlers/AutocompleteArgs.cs
@@ -1,0 +1,11 @@
+using Umea.se.EstateService.Shared.Autocomplete;
+
+namespace Umea.se.EstateService.Logic.Handlers;
+
+public sealed record AutocompleteArgs
+{
+    public AutocompleteType Type { get; init; } = AutocompleteType.Any;
+    public string Query { get; init; } = string.Empty;
+    public int Limit { get; init; } = 10;
+    public int? BuildingId { get; init; }
+}
