+++ src/ume-app-estateservice/Umea.se.EstateService.Logic/Handlers/AutocompleteHandler.cs
@@ -6,19 +6,6 @@ using Umea.se.EstateService.Shared.Autocomplete;
 
 namespace Umea.se.EstateService.Logic.Handlers;
 
-public interface IAutocompleteHandler
-{
-    Task<AutocompleteResult> SearchAsync(AutocompleteArgs request, CancellationToken cancellationToken);
-}
-
-public sealed record AutocompleteArgs
-{
-    public AutocompleteType Type { get; init; } = AutocompleteType.Any;
-    public string Query { get; init; } = string.Empty;
-    public int Limit { get; init; } = 10;
-    public int? BuildingId { get; init; }
-}
-
 public sealed class AutocompleteHandler(IPythagorasHandler pythagorasService, IMemoryCache cache) : IAutocompleteHandler
 {
     private static readonly TimeSpan _cacheDuration = TimeSpan.FromSeconds(30);
