+++ src/ume-app-estateservice/Umea.se.EstateService.Logic/Interfaces/IAutocompleteHandler.cs
@@ -1,0 +1,8 @@
+using Umea.se.EstateService.Logic.Handlers;
+
+namespace Umea.se.EstateService.Logic.Interfaces;
+
+public interface IAutocompleteHandler
+{
+    Task<AutocompleteResult> SearchAsync(AutocompleteArgs request, CancellationToken cancellationToken);
+}
