+++ src/ume-app-estateservice/Umea.se.EstateService.ServiceAccess/Pythagoras/Api/IPythagorasClient.cs
@@ -1,0 +1,7 @@
+namespace Umea.se.EstateService.ServiceAccess.Pythagoras.Api;
+
+public interface IPythagorasClient
+{
+    Task<IReadOnlyList<TDto>> GetAsync<TDto>(string endpoint, PythagorasQuery<TDto>? query, CancellationToken cancellationToken = default) where TDto : class;
+    IAsyncEnumerable<TDto> GetPaginatedAsync<TDto>(string endpoint, PythagorasQuery<TDto>? query, int pageSize = 50, CancellationToken cancellationToken = default) where TDto : class;
+}
