+++ src/ume-app-estateservice/Umea.se.EstateService.ServiceAccess/Pythagoras/Api/PythagorasClient.cs
@@ -3,12 +3,6 @@ using System.Text.Json;
 
 namespace Umea.se.EstateService.ServiceAccess.Pythagoras.Api;
 
-public interface IPythagorasClient
-{
-    Task<IReadOnlyList<TDto>> GetAsync<TDto>(string endpoint, PythagorasQuery<TDto>? query, CancellationToken cancellationToken = default) where TDto : class;
-    IAsyncEnumerable<TDto> GetPaginatedAsync<TDto>(string endpoint, PythagorasQuery<TDto>? query, int pageSize = 50, CancellationToken cancellationToken = default) where TDto : class;
-}
-
 public sealed class PythagorasClient(IHttpClientFactory httpClientFactory) : IPythagorasClient
 {
     private static readonly JsonSerializerOptions _serializerOptions = new(JsonSerializerDefaults.Web)
